refactored more station related funcs into it's module. added handpicking to handcoded stations. added a new method to eclipse map to map the final selected stations used by everything. can switch between selectedStations by swapping in one place at stations.js
