+++ src/client/eclipse-map.html
@@ -15,6 +15,7 @@
 <body>
     method:
     <select id="method">
+        <option value="selectedStationsJson">selectedStations</option>
         <option value="selectSectionBased">selectSectionBased</option>
         <option value="selectPointBased">selectPointBased</option>
         <option value="availableStationsJson">availableStations</option>
